# sap/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KomfortIntegration Documentation",
        "description": "OpenApi documentation for KomfortIntegration",
        "contact": {
            "email": "pawel@tikrow.com"
        },
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "KomfortIntegration API Server"
        }
    ],
    "paths": {
        "/data": {
            "post": {
                "summary": "Post data",
                "description": "Endpoint responsible for processing data",
                "operationId": "postData",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostDataRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request successful"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Payload": {
                "title": "Action payload",
                "description": "Action payload.",
                "required": [
                    "idUser",
                    "type",
                    "name",
                    "email",
                    "phoneNumber",
                    "isActive",
                    "idStore",
                    "city",
                    "address",
                    "postalCode",
                    "idMeasurementOrder",
                    "creationDateTime",
                    "installationCity",
                    "installationAddress",
                    "installationName",
                    "idAssembly",
                    "firstDisplayDateTime",
                    "userFirstDisplayDateTime",
                    "userAssignationDateTime",
                    "userDemissionDateTime"
                ],
                "properties": {
                    "idUser": {
                        "title": "Id User",
                        "type": "integer"
                    },
                    "type": {
                        "title": "Type",
                        "type": "string",
                        "nullable": true
                    },
                    "name": {
                        "title": "Name",
                        "type": "string",
                        "nullable": true
                    },
                    "firstName": {
                        "title": "First name",
                        "type": "string",
                        "nullable": true
                    },
                    "lastName": {
                        "title": "Last name",
                        "type": "string",
                        "nullable": true
                    },
                    "email": {
                        "title": "Email",
                        "type": "string",
                        "nullable": true
                    },
                    "npsValue": {
                        "title": "NPS value",
                        "type": "integer"
                    },
                    "dailyProductivity": {
                        "title": "Daily productivity",
                        "type": "integer"
                    },
                    "listOfCategory": {
                        "title": "List of category",
                        "type": "string",
                        "nullable": true
                    },
                    "phoneNumber": {
                        "title": "Number phone",
                        "type": "string",
                        "nullable": true
                    },
                    "isActive": {
                        "title": "Is active",
                        "type": "integer"
                    },
                    "idStore": {
                        "title": "Id Shop",
                        "type": "string",
                        "nullable": true
                    },
                    "city": {
                        "title": "City",
                        "type": "string",
                        "nullable": true
                    },
                    "address": {
                        "title": "Address",
                        "type": "string",
                        "nullable": true
                    },
                    "postalCode": {
                        "title": "Postal code",
                        "type": "string",
                        "nullable": true
                    },
                    "idMeasurementOrder": {
                        "title": "Id measurement order",
                        "type": "integer"
                    },
                    "idInstallationOrder": {
                        "title": "Id installation order",
                        "type": "integer"
                    },
                    "idParentInstallationOrder": {
                        "title": "Id parent installation order",
                        "type": "integer"
                    },
                    "measurementStatus": {
                        "title": "Measurement status",
                        "type": "string",
                        "nullable": true
                    },
                    "installationStatus": {
                        "title": "Installation status",
                        "type": "string",
                        "nullable": true
                    },
                    "assemblyCreatorName": {
                        "title": "Assembly creator name",
                        "type": "string",
                        "nullable": true
                    },
                    "creationDateTime": {
                        "title": "Creation Datetime",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updateDateTime": {
                        "title": "Update DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "deliveryDateTime": {
                        "title": "Delivery DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "expectedContactDateTime": {
                        "title": "Expected contact DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "expectedInstallationDateTime": {
                        "title": "Expected installation DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "acceptedInstallationDateTime": {
                        "title": "Accepted installation DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "floorCarpetMeters": {
                        "title": "Floor carpet meters",
                        "type": "integer"
                    },
                    "floorPanelMeters": {
                        "title": "Floor panel meters",
                        "type": "integer"
                    },
                    "floorWoodMeters": {
                        "title": "Floor wood meters",
                        "type": "integer"
                    },
                    "doorNumber": {
                        "title": "Door number",
                        "type": "integer"
                    },
                    "notificationEmail": {
                        "title": "Notification email",
                        "type": "string",
                        "nullable": true
                    },
                    "estimatedCostNet": {
                        "title": "Estimated cost net",
                        "type": "string",
                        "nullable": true
                    },
                    "installationCity": {
                        "title": "Installation city",
                        "type": "string",
                        "nullable": true
                    },
                    "installationAddress": {
                        "title": "Installation address",
                        "type": "string",
                        "nullable": true
                    },
                    "installationZipCode": {
                        "title": "Installation zip code",
                        "type": "string",
                        "nullable": true
                    },
                    "installationName": {
                        "title": "Installation name",
                        "type": "string",
                        "nullable": true
                    },
                    "installationPhoneNumber": {
                        "title": "Installation phone number",
                        "type": "string",
                        "nullable": true
                    },
                    "installationEmail": {
                        "title": "Installation email",
                        "type": "string",
                        "nullable": true
                    },
                    "installationNote": {
                        "title": "Installation note",
                        "type": "string",
                        "nullable": true
                    },
                    "idUserRanking": {
                        "title": "Id user ranking (array)",
                        "type": "string",
                        "nullable": true
                    },
                    "rankingPosition": {
                        "title": "Ranking position (array)",
                        "type": "string",
                        "nullable": true
                    },
                    "idAssembly": {
                        "title": "Id assembly",
                        "type": "integer"
                    },
                    "firstDisplayDateTime": {
                        "title": "First display DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "userFirstDisplayDateTime": {
                        "title": "User first display DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "userAssignationDateTime": {
                        "title": "User assignation DateTime",
                        "type": "string",
                        "nullable": true
                    },
                    "userDemissionDateTime": {
                        "title": "User demission DateTime",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PostDataRequest": {
                "title": "Post data request",
                "description": "Post data request body data",
                "required": [
                    "method",
                    "payload"
                ],
                "properties": {
                    "method": {
                        "title": "method",
                        "description": "API action to be executed",
                        "type": "string",
                        "example": "create_user|update_user|create_user_store|delete_user_store|create_store|update_store|create_assembly_order|update_assembly_order"
                    },
                    "payload": {
                        "$ref": "#/components/schemas/Payload"
                    }
                },
                "type": "object"
            }
        }
    }
}